Fix pavucontrol freezes
The volume control configured for i3 blocks has an icon and a text
script, which are separate. The icon was configured to call setsid,
which would run pavucontrol independently to the i3bar process. The
text script just ran the pavucontrol command normally without setsid.

Because i3bar is configured to hide itself automatically, and because
hiding the bar sends SIGSTOP to all children, pavucontrol would freeze
when run with the text command. The script has been fixes so
pavucontrol will always be called with setsid and no longer freeze